cleaning Code adding componetns
