+++ client/src/components/GFGFooter/index.js
@@ -1,0 +1,64 @@
+import React,{Component} from 'react'
+import {
+  Container,
+  Divider,
+  Dropdown,
+  Grid,
+  Header,
+  Image,
+  List,
+  Menu,
+  Segment,
+} from 'semantic-ui-react'
+
+class Footer extends Component {
+
+  render() {
+    const { mobile, isLoggedIn } = this.props;
+    return (
+      <Segment inverted vertical style={{ padding: '5em 0em' }}>
+      <Divider
+      as='h4'
+      className='header'
+      horizontal
+      style={{ margin: '3em 0em', textTransform: 'uppercase' }}
+    >
+      <a href='#'>Case Studies</a>
+    </Divider>
+        <Container>
+          <Grid divided inverted stackable>
+            <Grid.Row>
+              <Grid.Column width={3}>
+                <Header inverted as='h4' content='About' />
+                <List link inverted>
+                  <List.Item as='a'>Sitemap</List.Item>
+                  <List.Item as='a'>Contact Us</List.Item>
+                  <List.Item as='a'>Religious Ceremonies</List.Item>
+                  <List.Item as='a'>Gazebo Plans</List.Item>
+                </List>
+              </Grid.Column>
+              <Grid.Column width={3}>
+                <Header inverted as='h4' content='Services' />
+                <List link inverted>
+                  <List.Item as='a'>Banana Pre-Order</List.Item>
+                  <List.Item as='a'>DNA FAQ</List.Item>
+                  <List.Item as='a'>How To Access</List.Item>
+                  <List.Item as='a'>Favorite X-Men</List.Item>
+                </List>
+              </Grid.Column>
+              <Grid.Column width={7}>
+                <Header as='h4' inverted>
+                  Footer Header
+        </Header>
+                <p>
+                  Extra space for a call to action inside the footer that could help re-engage users.
+        </p>
+              </Grid.Column>
+            </Grid.Row>
+          </Grid>
+        </Container>
+      </Segment>
+    )
+  }
+}
+export default Footer
